+++ tests/Fixtures/TestBundle/Document/Issue7916/User.php
@@ -1,0 +1,66 @@
+<?php
+
+/*
+ * This file is part of the API Platform project.
+ *
+ * (c) Kévin Dunglas <dunglas@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace ApiPlatform\Tests\Fixtures\TestBundle\Document\Issue7916;
+
+use Doctrine\ODM\MongoDB\Mapping\Annotations as ODM;
+
+/**
+ * User document for issue #7916.
+ *
+ * This document is NOT marked with #[ApiResource] to test nested property filtering
+ * on relations to non-API-Resource entities.
+ *
+ * @see https://github.com/api-platform/core/issues/7916
+ */
+#[ODM\Document(collection: 'issue_7916_user')]
+class User
+{
+    #[ODM\Id(type: 'int', strategy: 'INCREMENT')]
+    private ?int $id = null;
+
+    #[ODM\Field(type: 'string')]
+    private ?string $name = null;
+
+    #[ODM\Field(type: 'string')]
+    private ?string $email = null;
+
+    public function getId(): ?int
+    {
+        return $this->id;
+    }
+
+    public function getName(): ?string
+    {
+        return $this->name;
+    }
+
+    public function setName(?string $name): self
+    {
+        $this->name = $name;
+
+        return $this;
+    }
+
+    public function getEmail(): ?string
+    {
+        return $this->email;
+    }
+
+    public function setEmail(?string $email): self
+    {
+        $this->email = $email;
+
+        return $this;
+    }
+}