+++ tests/Fixtures/TestBundle/Entity/Issue7916/UserAction.php
@@ -1,0 +1,70 @@
+<?php
+
+/*
+ * This file is part of the API Platform project.
+ *
+ * (c) Kévin Dunglas <dunglas@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace ApiPlatform\Tests\Fixtures\TestBundle\Entity\Issue7916;
+
+use Doctrine\ORM\Mapping as ORM;
+
+/**
+ * UserAction entity for issue #7916.
+ *
+ * This entity has a relation to User (which is NOT an ApiResource).
+ * Used to test nested property filtering on non-ApiResource relations with PartialSearchFilter.
+ *
+ * @see https://github.com/api-platform/core/issues/7916
+ */
+#[ORM\Entity]
+#[ORM\Table(name: 'issue_7916_user_action')]
+class UserAction
+{
+    #[ORM\Id]
+    #[ORM\GeneratedValue]
+    #[ORM\Column]
+    private ?int $id = null;
+
+    #[ORM\Column(length: 50)]
+    private string $action = '';
+
+    #[ORM\ManyToOne(targetEntity: User::class)]
+    #[ORM\JoinColumn(nullable: false)]
+    private ?User $user = null;
+
+    public function getId(): ?int
+    {
+        return $this->id;
+    }
+
+    public function getAction(): string
+    {
+        return $this->action;
+    }
+
+    public function setAction(string $action): self
+    {
+        $this->action = $action;
+
+        return $this;
+    }
+
+    public function getUser(): ?User
+    {
+        return $this->user;
+    }
+
+    public function setUser(?User $user): self
+    {
+        $this->user = $user;
+
+        return $this;
+    }
+}